refactoring: split handlers into several sub modules
