+++ vm_network_migration/handlers/backend_service_migration/backend_service_migration.py
@@ -18,8 +18,8 @@
 import warnings
 
 from vm_network_migration.errors import *
-from vm_network_migration.handlers.global_backend_service_migration import GlobalBackendServiceNetworkMigration
-from vm_network_migration.handlers.internal_backend_service_migration import InternalBackendServiceNetworkMigration
+from vm_network_migration.handlers.backend_service_migration.global_backend_service_migration import GlobalBackendServiceNetworkMigration
+from vm_network_migration.handlers.backend_service_migration.internal_backend_service_migration import InternalBackendServiceNetworkMigration
 from vm_network_migration.module_helpers.backend_service_helper import BackendServiceHelper
 from vm_network_migration.modules.backend_service_modules.backend_service import BackendService
 from vm_network_migration.modules.backend_service_modules.internal_backend_service import InternalBackendService
