+++ vm_network_migration/handlers/forwarding_rule_migration/external_forwarding_rule_migration.py
@@ -23,10 +23,10 @@ from vm_network_migration.handler_helper.selfLink_executor import SelfLinkExecut
 from vm_network_migration.module_helpers.forwarding_rule_helper import ForwardingRuleHelper
 from vm_network_migration.utils import initializer
 from vm_network_migration.handlers.compute_engine_resource_migration import ComputeEngineResourceMigration
-from vm_network_migration.handlers.backend_service_migration import BackendServiceMigration
+from vm_network_migration.handlers.backend_service_migration.backend_service_migration import BackendServiceMigration
 
 
-class InternalForwardingRuleMigration(ComputeEngineResourceMigration):
+class ExternalForwardingRuleMigration(ComputeEngineResourceMigration):
     @initializer
     def __init__(self, compute, project, forwarding_rule_name,
                  network_name, subnetwork_name,
@@ -42,7 +42,7 @@ class InternalForwardingRuleMigration(ComputeEngineResourceMigration):
             of the instances which serves this load balancer
             region: region of the internal load balancer
         """
-        super(InternalForwardingRuleMigration, self).__init__()
+        super(ExternalForwardingRuleMigration, self).__init__()
         self.forwarding_rule = self.build_forwarding_rule()
         self.backends_migration_handlers = []
 
@@ -61,38 +61,38 @@ class InternalForwardingRuleMigration(ComputeEngineResourceMigration):
         return forwarding_rule_helper.build_a_forwarding_rule()
 
 
+
     def network_migration(self):
-        """ Network migration for an internal forwarding rule.
-         The forwarding rule will be deleted first.
-         Then, the tool will migrate the backend service.
-         Finally, recreate the forwarding rule in the target subnet.
+        """ Network migration for a external forwarding rule.
+        The tool will migrate its backend services one by one.
+        The forwarding rule will not be deleted or recreated.
 
-         Returns:
-
-         """
+        """
         if self.forwarding_rule.compare_original_network_and_target_network():
             print('The backend service %s is already using target subnet.' % (
                 self.forwarding_rule_name))
             return
 
         backends_selfLinks = self.forwarding_rule.backends_selfLinks
+        if backends_selfLinks == []:
+            print(
+                'No backend service needs to be migrated. Terminating the migration.')
+            return
+
         for backends_selfLink in backends_selfLinks:
             selfLink_executor = SelfLinkExecutor(self.compute,
                                                  backends_selfLink,
                                                  self.network_name,
                                                  self.subnetwork_name,
                                                  self.preserve_instance_external_ip)
-            # the backends can be a target instance or an internal backend service
             try:
                 backends_migration_handler = selfLink_executor.build_migration_handler()
             except UnsupportedBackendService:
                 warnings.warn(
                     'The load balancing scheme of (%s) is not supported. '
                     'Continue migrating other backends.' % (backends_selfLink))
-                continue
+                continue            # Save handlers for rollback purpose
             if backends_migration_handler != None:
-                self.backends_migration_handlers.append(
-                    backends_migration_handler)
                 if isinstance(backends_migration_handler,
                               BackendServiceMigration):
                     backend_service = backends_migration_handler.backend_service
@@ -101,37 +101,16 @@ class InternalForwardingRuleMigration(ComputeEngineResourceMigration):
                             'The backend service is associated with two or more forwarding rules, \n'
                             'so it can not be migrated. \n'
                             'Terminating. ')
-                        return
-
-        print('Deleting: %s.' % (self.forwarding_rule_name))
-        self.forwarding_rule.delete_forwarding_rule()
-        for backends_migration_handler in self.backends_migration_handlers:
-            print('Migrating the backends of the forwarding rule.')
-            backends_migration_handler.network_migration()
-        print('Recreating the forwarding rule (%s) in the target subnet.' % (
-            self.forwarding_rule_name))
-        self.forwarding_rule.insert_forwarding_rule(
-            self.forwarding_rule.new_forwarding_rule_configs)
+                        # this backend service will be ignored and will continue migrate other backend services
+                        continue
+                self.backends_migration_handlers.append(
+                    backends_migration_handler)
+                backends_migration_handler.network_migration()
 
 
     def rollback(self):
-        """ Rollback an internal forwarding rule
-
-        Returns:
+        """ Error happens. Rollback to the original status.
 
         """
-        if not self.forwarding_rule.migrated and \
-                self.forwarding_rule.check_forwarding_rule_exists():
-            return
-        if self.forwarding_rule.migrated:
-            print('Deleting: %s.' %(self.forwarding_rule_name))
-            self.forwarding_rule.delete_forwarding_rule()
-            for backend_service_migration_handler in self.backends_migration_handlers:
-                backend_service_migration_handler.rollback()
-            self.forwarding_rule.insert_forwarding_rule(
-                self.forwarding_rule.forwarding_rule_configs)
-        else:
-            for backend_service_migration_handler in self.backends_migration_handlers:
-                backend_service_migration_handler.rollback()
-            self.forwarding_rule.insert_forwarding_rule(
-                self.forwarding_rule.forwarding_rule_configs)
+        for backend_service_migration_handler in self.backends_migration_handlers:
+            backend_service_migration_handler.rollback()
