+++ vm_network_migration/handlers/forwarding_rule_migration/forwarding_rule_migration.py
@@ -25,8 +25,8 @@ from vm_network_migration.modules.forwarding_rule_modules.external_regional_forw
 from vm_network_migration.modules.forwarding_rule_modules.external_global_forwarding_rule import ExternalGlobalForwardingRule
 from vm_network_migration.modules.forwarding_rule_modules.internal_regional_forwarding_rule import InternalRegionalForwardingRule
 from vm_network_migration.utils import initializer
-from vm_network_migration.handlers.external_forwarding_rule_migration import ExternalForwardingRuleMigration
-from vm_network_migration.handlers.internal_forwarding_rule_migration import InternalForwardingRuleMigration
+from vm_network_migration.handlers.forwarding_rule_migration.external_forwarding_rule_migration import ExternalForwardingRuleMigration
+from vm_network_migration.handlers.forwarding_rule_migration.internal_forwarding_rule_migration import InternalForwardingRuleMigration
 
 
 class ForwardingRuleMigration(ComputeEngineResourceMigration):
