+++ vm_network_migration/handlers/forwarding_rule_migration/internal_forwarding_rule_migration.py
@@ -23,10 +23,10 @@ from vm_network_migration.handler_helper.selfLink_executor import SelfLinkExecut
 from vm_network_migration.module_helpers.forwarding_rule_helper import ForwardingRuleHelper
 from vm_network_migration.utils import initializer
 from vm_network_migration.handlers.compute_engine_resource_migration import ComputeEngineResourceMigration
-from vm_network_migration.handlers.backend_service_migration import BackendServiceMigration
+from vm_network_migration.handlers.backend_service_migration.backend_service_migration import BackendServiceMigration
+from enum import IntEnum
 
-
-class ExternalForwardingRuleMigration(ComputeEngineResourceMigration):
+class InternalForwardingRuleMigration(ComputeEngineResourceMigration):
     @initializer
     def __init__(self, compute, project, forwarding_rule_name,
                  network_name, subnetwork_name,
@@ -42,9 +42,10 @@ class ExternalForwardingRuleMigration(ComputeEngineResourceMigration):
             of the instances which serves this load balancer
             region: region of the internal load balancer
         """
-        super(ExternalForwardingRuleMigration, self).__init__()
+        super(InternalForwardingRuleMigration, self).__init__()
         self.forwarding_rule = self.build_forwarding_rule()
         self.backends_migration_handlers = []
+        self.migration_status = MigrationStatus(0)
 
     def build_forwarding_rule(self):
         """ Use a helper class to create a ForwardingRule object
@@ -61,38 +62,38 @@ class ExternalForwardingRuleMigration(ComputeEngineResourceMigration):
         return forwarding_rule_helper.build_a_forwarding_rule()
 
 
-
     def network_migration(self):
-        """ Network migration for a external forwarding rule.
-        The tool will migrate its backend services one by one.
-        The forwarding rule will not be deleted or recreated.
+        """ Network migration for an internal forwarding rule.
+         The forwarding rule will be deleted first.
+         Then, the tool will migrate the backend service.
+         Finally, recreate the forwarding rule in the target subnet.
 
-        """
+         Returns:
+
+         """
         if self.forwarding_rule.compare_original_network_and_target_network():
             print('The backend service %s is already using target subnet.' % (
                 self.forwarding_rule_name))
             return
 
         backends_selfLinks = self.forwarding_rule.backends_selfLinks
-        if backends_selfLinks == []:
-            print(
-                'No backend service needs to be migrated. Terminating the migration.')
-            return
-
         for backends_selfLink in backends_selfLinks:
             selfLink_executor = SelfLinkExecutor(self.compute,
                                                  backends_selfLink,
                                                  self.network_name,
                                                  self.subnetwork_name,
                                                  self.preserve_instance_external_ip)
+            # the backends can be a target instance or an internal backend service
             try:
                 backends_migration_handler = selfLink_executor.build_migration_handler()
             except UnsupportedBackendService:
                 warnings.warn(
                     'The load balancing scheme of (%s) is not supported. '
                     'Continue migrating other backends.' % (backends_selfLink))
-                continue            # Save handlers for rollback purpose
+                continue
             if backends_migration_handler != None:
+                self.backends_migration_handlers.append(
+                    backends_migration_handler)
                 if isinstance(backends_migration_handler,
                               BackendServiceMigration):
                     backend_service = backends_migration_handler.backend_service
@@ -101,16 +102,48 @@ class ExternalForwardingRuleMigration(ComputeEngineResourceMigration):
                             'The backend service is associated with two or more forwarding rules, \n'
                             'so it can not be migrated. \n'
                             'Terminating. ')
-                        # this backend service will be ignored and will continue migrate other backend services
-                        continue
-                self.backends_migration_handlers.append(
-                    backends_migration_handler)
-                backends_migration_handler.network_migration()
+                        return
+        self.migration_status = MigrationStatus(1)
+        print('Deleting: %s.' % (self.forwarding_rule_name))
+        self.forwarding_rule.delete_forwarding_rule()
+        self.migration_status = MigrationStatus(2)
 
+        print('Migrating the backends of %s.' %(self.forwarding_rule_name))
+        for backends_migration_handler in self.backends_migration_handlers:
+            backends_migration_handler.network_migration()
+        self.migration_status = MigrationStatus(3)
+
+        print('Recreating the forwarding rule (%s) in the target subnet.' % (
+            self.forwarding_rule_name))
+        self.forwarding_rule.insert_forwarding_rule(
+            self.forwarding_rule.new_forwarding_rule_configs)
+        self.migration_status = MigrationStatus(4)
 
     def rollback(self):
-        """ Error happens. Rollback to the original status.
+        """ Rollback an internal forwarding rule
+
+        Returns:
 
         """
-        for backend_service_migration_handler in self.backends_migration_handlers:
-            backend_service_migration_handler.rollback()
+        if self.migration_status == 4:
+            print('Deleting: %s.' %(self.forwarding_rule_name))
+            self.forwarding_rule.delete_forwarding_rule()
+            self.migration_status = MigrationStatus(3)
+
+        if self.migration_status == 3:
+            for backend_service_migration_handler in self.backends_migration_handlers:
+                backend_service_migration_handler.rollback()
+            self.migration_status = MigrationStatus(2)
+
+        if self.migration_status == 2:
+            self.forwarding_rule.insert_forwarding_rule(
+                self.forwarding_rule.forwarding_rule_configs)
+            self.migration_status = MigrationStatus(0)
+
+
+class MigrationStatus(IntEnum):
+    NOT_START = 0
+    MIGRATING = 1
+    ORIGINAL_FORWARDING_RULE_DELETED = 2
+    BACKENDS_MIGRATED = 3
+    NEW_FORWARDING_RULE_CREATED = 4
