+++ vm_network_migration/handlers/instance_group_migration/instance_group_network_migration.py
@@ -24,8 +24,8 @@ from warnings import warn
 
 from vm_network_migration.errors import *
 from vm_network_migration.handlers.compute_engine_resource_migration import ComputeEngineResourceMigration
-from vm_network_migration.handlers.managed_instance_group_migration import ManagedInstanceGroupMigration
-from vm_network_migration.handlers.unmanaged_instance_group_migration import UnmanagedInstanceGroupMigration
+from vm_network_migration.handlers.instance_group_migration.managed_instance_group_migration import ManagedInstanceGroupMigration
+from vm_network_migration.handlers.instance_group_migration.unmanaged_instance_group_migration import UnmanagedInstanceGroupMigration
 from vm_network_migration.module_helpers.instance_group_helper import InstanceGroupHelper
 from vm_network_migration.modules.instance_group_modules.unmanaged_instance_group import UnmanagedInstanceGroup
 from vm_network_migration.utils import initializer
